requirements.txt
Django
django-cors-headers
djangorestframework
django-rest-swagger
psycopg2-binary
requests
django-filter
python-decouple
drf-yasg[validation]
django-phonenumber-field[phonenumbers]
Pillow
